+++ packages/create-analog/template-blog/src/app/pages/blog/[slug].page.ts
@@ -1,0 +1,29 @@
+import { Component } from '@angular/core';
+import { injectContent, MarkdownComponent } from '@analogjs/content';
+import { AsyncPipe } from '@angular/common';
+
+import PostAttributes from '../../post-attributes';
+
+@Component({
+  selector: 'app-blog-post',
+  standalone: true,
+  imports: [AsyncPipe, MarkdownComponent],
+  template: `
+    @if (post$ | async; as post) {
+    <article>
+      <img class="post__image" [src]="post.attributes.coverImage" />
+      <analog-markdown [content]="post.content" />
+    </article>
+    }
+  `,
+  styles: [
+    `
+      .post__image {
+        max-height: 40vh;
+      }
+    `,
+  ],
+})
+export default class HomeComponent {
+  readonly post$ = injectContent<PostAttributes>('slug');
+}